Address Copilot review fixes
