+++ pyproject.toml
@@ -10,7 +10,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "pyampp"
 version = "1.0.0"
-description = "Automatic Model Suff"
+description = "Python Automatic Model Production Pipeline"
 readme = "README.rst"
 requires-python = ">=3.10"
 license = { file = "licenses/LICENSE.rst", content-type = "text/plain" }
